+++ extensions/community-helper/lib/manifest.ts
@@ -1,6 +1,6 @@
 import { SERVER_HOST_PERMISSIONS } from "../../shared/constants";
 
-export const MANIFEST_PERMISSIONS = ["activeTab"] as const;
+export const MANIFEST_PERMISSIONS = ["storage", "activeTab"] as const;
 
 export const MANIFEST_HOST_PERMISSIONS = [...SERVER_HOST_PERMISSIONS] as const;